Add all tests to SignInTest
