+++ apps/account/tests/__init__.py
@@ -1,2 +1,1 @@
-from .authentication.manager import UsersManagersTests
-from .authentication.authenticate import AuthenticateTests
+from .authentication.auth import SignInTest
