--- apps/account/tests/authentication/authenticate.py
@@ -1,27 +1,0 @@
-from django.contrib.auth import get_user_model
-from django.test import TestCase, Client
-from ...models import UserManager
-
-
-class SignInTest(TestCase):
-
-    def setUp(self):
-        self.user = get_user_model().objects.create_user(username='test',
-                                                         password='12test12',
-                                                         email='test@example.com')
-        self.user.save()
-
-    def tearDown(self):
-        self.user.delete()
-
-    def test_correct(self):
-        user = authenticate(username='test', password='12test12')
-        self.assertTrue((user is not None) and user.is_authenticated)
-
-    def test_wrong_username(self):
-        user = authenticate(username='wrong', password='12test12')
-        self.assertFalse(user is not None and user.is_authenticated)
-
-    def test_wrong_pssword(self):
-        user = authenticate(username='test', password='wrong')
-        self.assertFalse(user is not None and user.is_authenticated)
